修改微信小程序配置实现类
